Initial port of Fabric version to MultiLoader Template
Basically copied everything from the Fabric only version to work with this

Will need to move some code to the common folder before porting to NeoForge
