+++ common/src/main/java/issame/material_beacons/Constants.java
@@ -1,0 +1,11 @@
+package issame.material_beacons;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Constants {
+
+	public static final String MOD_ID = "material_beacons";
+	public static final String MOD_NAME = "Material Beacons";
+	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+}
