+++ common/src/main/resources/assets/material_beacons/icon.png
