+++ common/src/main/resources/data/material_beacons/beacon/honey.json
