+++ common/src/main/resources/data/material_beacons/beacon/honeycomb.json
