+++ common/src/main/resources/data/material_beacons/beacon/iron.json
