+++ common/src/main/resources/data/material_beacons/beacon/redstone.json
