+++ fabric/src/main/java/issame/material_beacons/platform/FabricPlatformHelper.java
@@ -1,0 +1,24 @@
+package issame.material_beacons.platform;
+
+import issame.material_beacons.platform.services.IPlatformHelper;
+import net.fabricmc.loader.api.FabricLoader;
+
+public class FabricPlatformHelper implements IPlatformHelper {
+
+    @Override
+    public String getPlatformName() {
+        return "Fabric";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+
+        return FabricLoader.getInstance().isModLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+
+        return FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+}
