+++ gradle.properties
@@ -1,16 +1,39 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
+# Important Notes:
+# Every field you add must be added to buildSrc/src/main/groovy/multiloader-common.gradle expandProps map.
 
-# Fabric Properties
-	# check these on https://fabricmc.net/develop
-	minecraft_version=1.21.5
-	yarn_mappings=1.21.5+build.1
-	loader_version=0.16.10
+# Project
+version=1.1.0+1.21.6
+group=issame
+java_version=21
 
-# Mod Properties
-	mod_version = 1.0.1+1.21.5
-	maven_group = issame
-	archives_base_name = material-beacons
+# Common
+minecraft_version=1.21.5
+mod_name=Material Beacons
+mod_author=IssaMe
+mod_id=material_beacons
+license=CC0-1.0
+credits=
+description=Changes the effect of beacons to be based on the material used to construct them.
+minecraft_version_range=[1.21.5, 1.22)
+## This is the version of minecraft that the 'common' project uses, you can find a list of all versions here
+## https://projects.neoforged.net/neoforged/neoform
+neo_form_version=1.21.5-20250325.162830
+# The version of ParchmentMC that is used, see https://parchmentmc.org/docs/getting-started#choose-a-version for new versions
+parchment_minecraft=1.21.5
+parchment_version=2025.04.19
 
-# Dependencies
-	fabric_version=0.119.6+1.21.5
+# Fabric
+fabric_version=0.124.0+1.21.5
+fabric_loader_version=0.16.14
+
+# Forge
+forge_version=55.0.1
+forge_loader_version_range=[55,)
+
+# NeoForge
+neoforge_version=21.5.66-beta
+neoforge_loader_version_range=[4,)
+
+# Gradle
+org.gradle.jvmargs=-Xmx3G
+org.gradle.daemon=false
