+++ neoforge/src/main/java/issame/material_beacons/ExampleMod.java
@@ -1,0 +1,21 @@
+package issame.material_beacons;
+
+
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.common.Mod;
+
+@Mod(Constants.MOD_ID)
+public class ExampleMod {
+
+    public ExampleMod(IEventBus eventBus) {
+
+        // This method is invoked by the NeoForge mod loader when it is ready
+        // to load your mod. You can access NeoForge and Common code in this
+        // project.
+
+        // Use NeoForge to bootstrap the Common mod.
+        Constants.LOG.info("Hello NeoForge world!");
+        CommonClass.init();
+
+    }
+}
