+++ neoforge/src/main/java/issame/material_beacons/mixin/MixinTitleScreen.java
@@ -1,0 +1,20 @@
+package issame.material_beacons.mixin;
+
+import issame.material_beacons.Constants;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.TitleScreen;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(TitleScreen.class)
+public class MixinTitleScreen {
+
+    @Inject(at = @At("HEAD"), method = "init()V")
+    private void init(CallbackInfo info) {
+
+        Constants.LOG.info("This line is printed by an example mod mixin from NeoForge!");
+        Constants.LOG.info("MC Version: {}", Minecraft.getInstance().getVersionType());
+    }
+}
