+++ neoforge/src/main/java/issame/material_beacons/platform/NeoForgePlatformHelper.java
@@ -1,0 +1,26 @@
+package issame.material_beacons.platform;
+
+import issame.material_beacons.platform.services.IPlatformHelper;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.loading.FMLLoader;
+
+public class NeoForgePlatformHelper implements IPlatformHelper {
+
+    @Override
+    public String getPlatformName() {
+
+        return "NeoForge";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+
+        return ModList.get().isLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+
+        return !FMLLoader.isProduction();
+    }
+}
