+++ neoforge/src/main/resources/META-INF/services/issame.material_beacons.platform.services.IPlatformHelper
@@ -1,0 +1,1 @@
+issame.material_beacons.platform.NeoForgePlatformHelper
