--- src/main/resources/fabric.mod.json
@@ -1,35 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "material_beacons",
-  "version": "${version}",
-
-  "name": "Material Beacons",
-  "description": "Changes the effect of beacons to be based on the material used to construct them.",
-  "authors": [
-    "IssaMe"
-  ],
-  "contact": {
-    "homepage": "https://fabricmc.net/",
-    "sources": "https://github.com/FabricMC/fabric-example-mod"
-  },
-
-  "license": "MIT",
-  "icon": "assets/material_beacons/icon.png",
-
-  "environment": "*",
-  "entrypoints": {
-    "main": [
-      "issame.material_beacons.MaterialBeacons"
-    ]
-  },
-  "mixins": [
-    "material_beacons.mixins.json"
-  ],
-
-  "depends": {
-    "fabricloader": ">=0.16.10",
-    "fabric-api": "*",
-    "minecraft": "~1.21.5",
-    "java": ">=21"
-  }
-}
