--- src/main/resources/material_beacons.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "issame.material_beacons.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "mixins": [
-    "BeaconBlockEntityMixin"
-  ],
-  "client": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
